// src/shared/type.enum.ts
export enum Type  {
    bug = 0,
    dark = 1,
    dragon = 2,
    electric = 3,
    fairy = 4,
    fighting = 5,
    fire = 6,
    flying = 7,
    ghost = 8,
    grass = 9,
    ground = 10,
    ice = 11,
    normal = 12,
    poison = 13,
    psychic = 14,
    rock = 15,
    steel = 16,
    water = 17,
    unknown = 18,
}